requirements.txt
Flask
pyclamd
python-pptx
PyPDF2
olefile
Pillow
python-docx
